[6/11] Refactor index controller tests
Collect tests covering index controller in a single module
